Move websocket tests to a separate file
